Remove building of shared library for aiebu
We want all clients to link statically with aiebu and now build
and release only libaiebu_static.a (linux) and aiebu_static.lib (windows)

Import and export macros (dllimport/export) were not set up correctly
on windows, building only the static library is simpler as it avoids
import/export completely.

Revisit when/if there is a need for shared library.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>