Remove _static suffix from static libraries on Linux
Static libraries are named same as shared libraries.  CMake targets
are unchanged to distinguish linking with static vs shared, but refer
to the renamed static library.

Sample CMakeLists.txt change:
```
if (NOT MSVC)
  # On linux use xrt_coreutil.a
  set_target_properties(aiebu_static PROPERTIES OUTPUT_NAME "aiebu")
endif()
```
Both the build and install version of the static libraries are changed, but
internal build targets that link against static libraries continue to refer
to e.g. `aiebu_static`. Under the hood CMake is aware of the renaming.

While AIEBU only builds a static library, I chose to follow the XRT
convention to support both static and shared just in case we add a
shared library later.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>