spec/build: Don't specify version controlled header as BYPRODUCTS.
This resolves an issue of the isa.h header being incorrectly regenerated
when invoking the build a second time using the Ninja generator.

This resulted in a broken build due to isa_disassembler being
undefined in the regenerated header.

Due to differences in CMake generators regarding to BYPRODUCS,
the issue was not reproducible using the Unix Makefile generator

https://cmake.org/cmake/help/latest/command/add_custom_target.html

In general BYPRODUCTS is not intended to overwrite version controlled
source files.

This fixes issue #144.